+++ lib/rules/one_of_rule.dart
@@ -1,15 +1,15 @@
 import 'package:flrx_validator/rule.dart';
+import 'package:flrx_validator/rules/in_rule.dart';
 
 /// A [Rule] subclass validating if the input is one of the element in the [acceptedList].
-class OneOfRule<T> extends Rule<T> {
-  final List<T> acceptedList;
+@Deprecated('Use InRule instead')
+class OneOfRule<T> extends InRule<T> {
 
-  OneOfRule(this.acceptedList, {String validationMessage})
-      : super(validationMessage);
+  OneOfRule(List<T> acceptedList, {String validationMessage})
+      : super(acceptedList, validationMessage: validationMessage);
 
   @override
   String onValidate(String entityName, T value) {
-    String validationMessage = ":entity is not in list of accepted values";
-    return acceptedList.contains(value) ? null : validationMessage;
+    return super.onValidate(entityName, value);
   }
 }