+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [v0.3.0]
+
+### Added
+- InRule, NotInRule
+
+### Deprecated
+- OneOfRule in favour of InRule
+
+
+### Changed
+- Minor Code Cleanup.
+
 ## [v0.2.0]
 
 ### Added